explorer/requirements.txt
requests~=2.31.0
pytz~=2023.3.post1
Flask~=3.0.0
Flask-Limiter~=2.1
json2html~=1.3.0
Werkzeug~=3.0.0
numpy~=1.22.1
networkx~=2.6.3
gunicorn==20.1.0
pycryptodome==3.14.1
psycopg2-binary==2.9.6
flask-restx~=1.3.0
scipy~=1.10.1
flask-turnstile~=0.1.1
pymemcache~=4.0.0